Normalize alt as uint16 t 2

+++ m20/Core/Inc/xm_gps.h
@@ -11,7 +11,7 @@ typedef struct TXMDATA {
       Fix; // 00 = no data/parsing error, 01 = no fix, 02 = 2D fix, 03 = 3D fix
   float Lat;        // latitude in degrees with decimal places + for N - for S
   float Lon;        // longitude in degrees with decimal places
-  float Alt;        // altitude in meters
+  uint16_t Alt;     // altitude in meters
   float AscentRate; // m/s
   float GroundSpeed;
   uint32_t Time;
